+++ app/src/main/java/com/battlelancer/seriesguide/SgApp.kt
@@ -10,7 +10,6 @@ import android.os.StrictMode
 import android.os.StrictMode.ThreadPolicy
 import android.os.StrictMode.VmPolicy
 import androidx.annotation.RequiresApi
-import com.battlelancer.seriesguide.extensions.ExtensionManager
 import com.battlelancer.seriesguide.modules.AppModule
 import com.battlelancer.seriesguide.modules.DaggerServicesComponent
 import com.battlelancer.seriesguide.modules.HttpClientModule
@@ -32,6 +31,7 @@ import io.palaima.debugdrawer.timber.data.LumberYard
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.newSingleThreadContext
 import org.greenrobot.eventbus.EventBus
 import org.greenrobot.eventbus.EventBusException
 import timber.log.Timber
@@ -97,6 +97,10 @@ class SgApp : Application() {
          */
         const val CONTENT_AUTHORITY = BuildConfig.APPLICATION_ID + ".provider"
 
+        /** Executes one coroutine at a time. But does not guarantee order if they suspend. */
+        @Suppress("EXPERIMENTAL_API_USAGE")
+        val SINGLE = newSingleThreadContext("SingleThread")
+
         private var servicesComponent: ServicesComponent? = null
 
         @JvmStatic
@@ -138,8 +142,6 @@ class SgApp : Application() {
 
         // Load the current theme into a global variable
         ThemeUtils.updateTheme(DisplaySettings.getThemeIndex(this))
-
-        ExtensionManager.get().checkEnabledExtensions(this)
 
         // Tell Google Play Services to update the security provider.
         // This is not technically required, but might improve connection issues